Prevent OSError exception in unary_union()
This issue was already discussed in #107 and fixed for connect_points.py
in b25921f. Now the same access violation error came up at this point.
It could be confirmed that the issue happens only with shapely < 2.0
and is not occuring with shapely >= 2.0. If that ever becomes a
requirement, the 'list()' could be removed.